Rename s3caller to units3caller
We'll require having multiple types of s3caller workers, so we
need to namespace them to prevent collisions. Also I don't want
to make the unit s3 caller more complicated and harder to reason
about.
