+++ internal/worker/units3caller/api_mocks.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -package s3caller -destination api_mocks.go github.com/juju/juju/api Connection
+//	mockgen -package units3caller -destination api_mocks.go github.com/juju/juju/api Connection
 //
 
-// Package s3caller is a generated GoMock package.
-package s3caller
+// Package units3caller is a generated GoMock package.
+package units3caller
 
 import (
 	context "context"
