+++ internal/worker/units3caller/package_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -package s3caller -destination package_mock_test.go github.com/juju/juju/core/objectstore Session
+//	mockgen -package units3caller -destination package_mock_test.go github.com/juju/juju/core/objectstore Session
 //
 
-// Package s3caller is a generated GoMock package.
-package s3caller
+// Package units3caller is a generated GoMock package.
+package units3caller
 
 import (
 	context "context"
