+++ internal/worker/units3caller/package_test.go
@@ -1,7 +1,7 @@
 // Copyright 2023 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package s3caller
+package units3caller
 
 import (
 	"testing"
@@ -13,8 +13,8 @@ import (
 	coretesting "github.com/juju/juju/testing"
 )
 
-//go:generate go run go.uber.org/mock/mockgen -package s3caller -destination package_mock_test.go github.com/juju/juju/core/objectstore Session
-//go:generate go run go.uber.org/mock/mockgen -package s3caller -destination api_mocks.go github.com/juju/juju/api Connection
+//go:generate go run go.uber.org/mock/mockgen -package units3caller -destination package_mock_test.go github.com/juju/juju/core/objectstore Session
+//go:generate go run go.uber.org/mock/mockgen -package units3caller -destination api_mocks.go github.com/juju/juju/api Connection
 
 func TestPackage(t *testing.T) {
 	gc.TestingT(t)
