+++ internal/worker/units3caller/worker.go
@@ -1,7 +1,7 @@
 // Copyright 2023 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package s3caller
+package units3caller
 
 import (
 	"github.com/juju/worker/v4"
